+++ code/src/cljs/sixsq/nuvla/ui/deployments/views.cljs
@@ -40,7 +40,8 @@
         [full-text-search-plugin/FullTextSearch
          {:db-path      [::spec/deployments-search]
           :change-event [::pagination-plugin/change-page
-                         [::spec/pagination] 1]}]
+                         [::spec/pagination] 1]
+          :placeholder-suffix @(subscribe [::subs/state-selector])}]
         " "
         ^{:key (random-uuid)}
         [filter-comp/ButtonFilter